+++ lahja/base.py
@@ -23,12 +23,15 @@ from typing import (  # noqa: F401
 from .common import (
     BaseEvent,
     BaseRequestResponseEvent,
+    Broadcast,
     BroadcastConfig,
     ConnectionConfig,
     Message,
     Msg,
     Subscription,
+    SubscriptionsUpdated,
 )
+from .exceptions import RemoteDisconnected
 from .typing import ConditionAPI, EventAPI, LockAPI
 
 TResponse = TypeVar("TResponse", bound=BaseEvent)
@@ -85,6 +88,9 @@ class RemoteEndpointAPI(ABC):
 
     _subscriptions_initialized: EventAPI
 
+    #
+    # Object run lifecycle
+    #
     @abstractmethod
     async def wait_started(self) -> None:
         ...
@@ -94,11 +100,11 @@ class RemoteEndpointAPI(ABC):
         ...
 
     @abstractmethod
-    async def is_running(self) -> bool:
+    def is_running(self) -> bool:
         ...
 
     @abstractmethod
-    async def is_stopped(self) -> bool:
+    def is_stopped(self) -> bool:
         ...
 
     @abstractmethod
@@ -113,6 +119,10 @@ class RemoteEndpointAPI(ABC):
     async def _run(self) -> None:
         ...
 
+    #
+    # Core external API
+    #
+    @abstractmethod
     async def notify_subscriptions_updated(
         self, subscriptions: Set[Type[BaseEvent]], block: bool = True
     ) -> None:
@@ -125,14 +135,103 @@ class RemoteEndpointAPI(ABC):
         """
         ...
 
+    @abstractmethod
     def can_send_item(self, item: BaseEvent, config: Optional[BroadcastConfig]) -> bool:
         ...
 
+    @abstractmethod
     async def send_message(self, message: Msg) -> None:
         ...
 
+    @abstractmethod
     async def wait_until_subscription_received(self) -> None:
         ...
+
+
+class BaseRemoteEndpoint(RemoteEndpointAPI):
+    """
+    This base class implements common logic that can be shared across different
+    implementations of the `RemoteEndpointAPI`
+
+    Represents a connection to another endpoint.  Connections *can* be
+    bi-directional with messages flowing in either direction.
+    """
+
+    logger = logging.getLogger("lahja.endpoint.asyncio.RemoteEndpoint")
+
+    def __init__(
+        self,
+        name: Optional[str],
+        conn: ConnectionAPI,
+        new_msg_func: Callable[[Broadcast], Awaitable[Any]],
+    ) -> None:
+        self.name = name
+        self.conn = conn
+        self.new_msg_func = new_msg_func
+
+        self.subscribed_messages: Set[Type[BaseEvent]] = set()
+
+    def __str__(self) -> str:
+        return f"RemoteEndpoint[{self.name if self.name is not None else id(self)}]"
+
+    def __repr__(self) -> str:
+        return f"<{self}>"
+
+    async def wait_started(self) -> None:
+        await self._running.wait()
+
+    async def wait_stopped(self) -> None:
+        await self._stopped.wait()
+
+    def is_running(self) -> bool:
+        return not self.is_stopped and self.running.is_set()
+
+    def is_stopped(self) -> bool:
+        return self._stopped.is_set()
+
+    async def notify_subscriptions_updated(
+        self, subscriptions: Set[Type[BaseEvent]], block: bool = True
+    ) -> None:
+        """
+        Alert the endpoint on the other side of this connection that the local
+        subscriptions have changed. If ``block`` is ``True`` then this function
+        will block until the remote endpoint has acknowledged the new
+        subscription set. If ``block`` is ``False`` then this function will
+        return immediately after the send finishes.
+        """
+        # The extra lock ensures only one coroutine can notify this endpoint at any one time
+        # and that no replies are accidentally received by the wrong
+        # coroutines. Without this, in the case where `block=True`, this inner
+        # block would release the lock on the call to `wait()` which would
+        # allow the ack from a different update to incorrectly result in this
+        # returning before the ack had been received.
+        async with self._notify_lock:
+            async with self._received_response:
+                try:
+                    await self.conn.send_message(
+                        SubscriptionsUpdated(subscriptions, block)
+                    )
+                except RemoteDisconnected:
+                    return
+                if block:
+                    await self._received_response.wait()
+
+    def can_send_item(self, item: BaseEvent, config: Optional[BroadcastConfig]) -> bool:
+        if config is not None:
+            if self.name is not None and not config.allowed_to_receive(self.name):
+                return False
+            elif config.filter_event_id is not None:
+                # the item is a response to a request.
+                return True
+
+        return type(item) in self.subscribed_messages
+
+    async def send_message(self, message: Msg) -> None:
+        await self.conn.send_message(message)
+
+    async def wait_until_subscription_received(self) -> None:
+        async with self._received_subscription:
+            await self._received_subscription.wait()
 
 
 class EndpointAPI(ABC):
